fix: linter and tests
