+++ pkg/webhooks/resource/fake.go
@@ -19,6 +19,7 @@ import (
 	"github.com/kyverno/kyverno/pkg/metrics"
 	"github.com/kyverno/kyverno/pkg/policycache"
 	"github.com/kyverno/kyverno/pkg/registryclient"
+	"github.com/kyverno/kyverno/pkg/utils/report"
 	"github.com/kyverno/kyverno/pkg/webhooks/updaterequest"
 	webhookutils "github.com/kyverno/kyverno/pkg/webhooks/utils"
 	kubeinformers "k8s.io/client-go/informers"
@@ -45,16 +46,17 @@ func NewFakeHandlers(ctx context.Context, policyCache policycache.Cache) *resour
 	rclient := registryclient.NewOrDie()
 
 	return &resourceHandlers{
-		client:        dclient,
-		configuration: configuration,
-		metricsConfig: metricsConfig,
-		pCache:        policyCache,
-		nsLister:      informers.Core().V1().Namespaces().Lister(),
-		urLister:      urLister,
-		urGenerator:   updaterequest.NewFake(),
-		eventGen:      event.NewFake(),
-		pcBuilder:     webhookutils.NewPolicyContextBuilder(configuration, jp),
-		auditPool:     pond.New(8, 1000),
+		client:          dclient,
+		configuration:   configuration,
+		metricsConfig:   metricsConfig,
+		pCache:          policyCache,
+		nsLister:        informers.Core().V1().Namespaces().Lister(),
+		urLister:        urLister,
+		urGenerator:     updaterequest.NewFake(),
+		eventGen:        event.NewFake(),
+		pcBuilder:       webhookutils.NewPolicyContextBuilder(configuration, jp),
+		auditPool:       pond.New(8, 1000),
+		reportingConfig: report.NewReportingConfig("validate", "mutate", "mutateExisiting", "generate", "imageVerify"),
 		engine: engine.NewEngine(
 			configuration,
 			config.NewDefaultMetricsConfiguration(),
